nie wiadomo

+++ database/migrations/2022_05_14_212235_create_profiles_table.php
@@ -20,7 +20,7 @@ class CreateProfilesTable extends Migration
             $table->string('image')->nullable();
             $table->string('town')->nullable();
             $table->string('street')->nullable();
-            $table->bigInteger('houseNumber')->nullable();
+            $table->tinyInteger('houseNumber')->nullable();
             $table->string('zipcode')->nullable();
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
